+++ src/main/java/ch/salon/service/ExhibitorService.java
@@ -1,7 +1,7 @@
 package ch.salon.service;
 
+import ch.salon.domain.Exhibitor;
 import ch.salon.domain.enumeration.EntityType;
-import ch.salon.repository.EventLogRepository;
 import ch.salon.repository.ExhibitorRepository;
 import ch.salon.service.dto.EventLogDTO;
 import ch.salon.service.dto.ExhibitorDTO;
@@ -10,6 +10,7 @@ import ch.salon.service.mapper.ExhibitorMapper;
 import ch.salon.web.rest.errors.BadRequestAlertException;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -21,28 +22,32 @@ public class ExhibitorService {
     public static final String ENTITY_NAME = "exhibitor";
 
     private final ExhibitorRepository exhibitorRepository;
-    private final EventLogRepository eventLogRepository;
     private final EventLogService eventLogService;
 
-    public ExhibitorService(ExhibitorRepository exhibitorRepository, EventLogRepository eventLogRepository,
-                            EventLogService eventLogService) {
+    public ExhibitorService(ExhibitorRepository exhibitorRepository, EventLogService eventLogService) {
         this.exhibitorRepository = exhibitorRepository;
-        this.eventLogRepository = eventLogRepository;
         this.eventLogService = eventLogService;
     }
 
     public UUID create(ExhibitorDTO exhibitor) {
+        if (exhibitor == null) {
+            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+
         if (exhibitor.getId() != null) {
             throw new BadRequestAlertException("A new exhibitor cannot already have an ID", ENTITY_NAME, "idexists");
         }
+        Exhibitor entity = ExhibitorMapper.INSTANCE.toEntity(exhibitor);
+        entity.setRegistrationDate(Instant.now());
 
-        return exhibitorRepository.save(ExhibitorMapper.INSTANCE.toEntity(exhibitor)).getId();
+        return exhibitorRepository.save(entity).getId();
     }
 
     public ExhibitorDTO update(final UUID id, ExhibitorDTO exhibitor) {
-        if (exhibitor.getId() == null) {
+        if (exhibitor == null || exhibitor.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+
         if (!Objects.equals(id, exhibitor.getId())) {
             throw new BadRequestAlertException("Invalid ID", ENTITY_NAME, "idinvalid");
         }
@@ -55,18 +60,33 @@ public class ExhibitorService {
     }
 
     public List<ExhibitorDTO> findAll() {
-        return exhibitorRepository.findAll().stream().map(ExhibitorMapper.INSTANCE::toDto).toList();
+        return exhibitorRepository.findByOrderByRegistrationDateDesc()
+                                  .stream()
+                                  .map(ExhibitorMapper.INSTANCE::toDto)
+                                  .toList();
     }
 
     public Optional<ExhibitorDTO> get(UUID id) {
+        if (id == null) {
+            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+
         return exhibitorRepository.findById(id).map(ExhibitorMapper.INSTANCE::toDto);
     }
 
     public void delete(UUID id) {
+        if (id == null) {
+            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+
         exhibitorRepository.deleteById(id);
     }
 
     public List<EventLogDTO> findAllEventLogs(UUID idExhibitor) {
+        if (idExhibitor == null) {
+            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+
         return this.eventLogService.findAllEventLog(EntityType.EXHIBITOR, idExhibitor)
                                    .stream()
                                    .map(EventLogMapper.INSTANCE::toDto)
